# docs/openapi/babil-public-api.yaml
openapi: 3.0.3
info:
  title: Babil public HTTP API (extract)
  description: |
    Maintained contract for the main **anonymous** JSON endpoints used by the app and integrators.
    Country records are large and evolve; schemas stay intentionally loose (`additionalProperties`) except where validated (engine POST bodies).

    **Live spec:** `GET /api/openapi` serves this file (`application/yaml`).
  version: 1.0.0
servers:
  - url: /
    description: Same origin as the Next.js deployment
...
paths:
...
  /api/recommendation:
    post:
      tags: [engines]
      summary: Recommendation engine (ranked countries for profile)
      operationId: postRecommendation
      parameters: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoProbaPostBody'
      responses:
        '200':
          description: Engine result (shape varies; includes `engineVersion` and related fields).
          headers:
            X-Babil-Engine-Version:
              schema: { type: string }
            X-Babil-Engine-Kind:
              schema: { type: string, example: recommendation }
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid JSON or Zod validation failure
        '413':
          description: Body too large
        '429':
          description: Rate limited
...
components:
  schemas:
...
    RecoProbaPostBody:
      type: object
      description: Validated subset; unknown top-level keys are accepted for forward compatibility.
      properties:
        profile:
          type: object
          additionalProperties: true
        playground:
          type: boolean
        focusCountryId:
          type: integer
          minimum: 1
          description: Optional country id to list first in ranked results (same public id as /countries/[id]).
      additionalProperties: true